Refactor barplot code.
